requirements.txt
pytest
requests
voluptuous
pytest_voluptuous
python-dotenv
curlify
allure-pytest
